Kill nrepl server process with SIGTERM
on all platforms. Originally, we killed it with SIGKILL. Then in 2017,
we added `nrepl--kill-process` to use SIGINT on Windows but kept using
SIGKILL for Linux and MacOS. This commit changes
`nrepl-kill-server-buffer` to send SIGTERM in all cases.
The problem with SIGKILL is it aborts the JVM too aggressively, before
shutdown hooks and other cleanup e.g. (.deleteOnExit ^java.io.File f)
have a chance to run.

TODO test on Windows and MacOS.